Port jquery-ui/shake to webpack and remove the patches
